refactor: restructure User module's mappers and DTOs
- Merge UserRequestMapper and UserResponseMapper into a single UserMapper interface
- Move DTOs from 'dtos' to 'dto' package for consistency
- Update service and controller classes to use new mapper
- Clean up imports and code formatting
- Replace LocalDateTime with ZonedDateTime for timestamps
