+++ user/src/main/java/com/fernirx/lms/user/mapper/UserMapper.java
@@ -1,14 +1,16 @@
 package com.fernirx.lms.user.mapper;
 
-import com.fernirx.lms.user.dtos.response.UserResponseDTO;
+
+import com.fernirx.lms.user.dto.request.UserRequestDTO;
+import com.fernirx.lms.user.dto.response.UserResponseDTO;
 import com.fernirx.lms.user.entity.User;
 import org.mapstruct.Mapper;
 
 import java.util.List;
 
 @Mapper(componentModel = "spring")
-public interface UserResponseMapper {
+public interface UserMapper {
+    User toEntity(UserRequestDTO user);
     UserResponseDTO toDto(User user);
-    User toEntity(UserResponseDTO userDTO);
     List<UserResponseDTO> toListDto(List<User> user);
 }
